refactor: finish private/features catalog refactor
