+++ src/app/features/private/group-preview/components/group-shop-view/group-item/group-item.component.css
