+++ src/app/features/private/group-preview/components/group-shop-view/group-shop/group-shop.component.css
