+++ src/app/features/private/group-preview/group-preview-page/group-preview-page.component.html
@@ -145,11 +145,11 @@
           <span nz-icon nzType="arrow-left"></span>
         </button>
       </div>
-      <app-group-shop-page
+      <app-group-shop
         [group]="group()!"
         [viewMode]="mode()"
         (deactivated)="closeShop()"
-      ></app-group-shop-page>
+      ></app-group-shop>
     </div>
   </nz-spin>
 </div>
