+++ src/app/features/private/group-ranking/components/full-ranking/full-ranking.component.html
