+++ src/app/features/private/group-ranking/group-ranking-page/group-ranking-page.component.css
